Disable example, xrandr-indicator, auto-move-windows, user-theme by default.
example: not really useful for the general public.
xrandr-indicator: requires some features in gjs and gobject-introspection
that are not merged yet (bugs 634253 and 643620).
auto-move-windows and user-theme: these are very useful, but require
GSettings, so can only be installed in the same prefix as GLib, or
otherwise one needs to set GSETTINGS_SCHEMA_DIR in .profile)