Auto Move Windows: open overview when last window is closed
Using the auto move window the overview is not shown when the last
window on a workspace is closed, instead the workspace before is
activated. Fix the extension to show the ovreview instead of changing
workspace when the last window is closed. The fix is basically merged
from gnome-shell master.

https://bugzilla.gnome.org/show_bug.cgi?id=651586