xrandr-indicator: don't reference a non existing property
I was planning to add global.stage_xwindow, back when this was
written for core. Then this became an extension, but the reference
staid there.
We could use clutter_x11_get_stage_window(), if that wasn't marked
(skip). For now, let's just 0 (it is correct anyway, as we want
the dialog to be system modal)