Convenience: relicense to BSD
This module is meant to be a copy-lib, and it is used all over the
place (including GPLv3, LGPL or BSD licensed extensions). I don't
want to block an extension for license incompatibility.

The rest of the repository is still GPLv2, as it is the tarball
distribution and individual extensions downloaded from the website.