auto-move-windows: allow specifying a monitor index too
In addition to force windows on specific workspaces, it is useful to
assign them a specific monitor too.
Allow to do so by extending the configuration format.

As a small change, windows now get moved to the primary monitor if
they match a configured app. This is in line with the idea that
workspaces are only on the primary monitor.

Fixes https://gitlab.gnome.org/GNOME/gnome-shell-extensions/issues/3