window-list: Set metaWindow for single-window items when grouped
When grouping is enabled, items that are only have a single window
associated are still expected to behave like ungrouped items.
This includes activating windows by DND, which uses the metaWindow
property on the hovered actor's _delegate, so make sure it is set
on AppButtons in single-window mode as well.

https://bugzilla.gnome.org/show_bug.cgi?id=693561