window-list: Restrict right-click actions to windows in list
While the list of windows in the left-click menu is filtered by
workspace, the minimize/maximize/close actions in the right-click
menu apply to all application windows on all workspaces.
This is fairly confusing, so restrict the actions to only apply
to windows that do appear in the left-click list.

https://bugzilla.gnome.org/show_bug.cgi?id=724134