window-list: Reposition on height changes
The window list position depends on both the monitor geometry and the
list height, however changes to the latter are currently ignored. For
the time being this doesn't matter due to the list's fixed height, but
we are about to scale the list with the text, so reposition the list
on height changes.

https://bugzilla.gnome.org/show_bug.cgi?id=703585