screenshot-window-sizer: hidpi support
The new logical dimensions are reported in the overlay, rather than the
pixel dimensions. That is: if scaleFactor is 2, a window might be
resized to 2400×1350 device pixels, which will be reported as 1200×675
in the overlay.

This is consistent with (for example) the DevTools in Chrome, which
reports the logical size of the viewport when you resize the window,
rather than the physical pixel size.

Tested with a freely-resizable window and with a constrained-geometry
window (GNOME Terminal), on a hidpi display.

https://bugzilla.gnome.org/show_bug.cgi?id=754607