places-menu: enable to unmount places
This patch enable to unmount removable devices and network shares
from the places menu by checking for each menu entry if the entry
is removable. If yes, an eject button is placed next to the entry
label and bind to an eject procedure.

Fixes https://gitlab.gnome.org/GNOME/gnome-shell-extensions/issues/17