build: Restrict enabled extensions to available ones
meson 0.44 features a new option type called `array` that allows
more than one string to be passed that are restricted with a set
of specified choices.

This feature fits perfectly the `enable_extensions` option needs,
so it has been changed to be an `array` type, and possible choices
have been set to the available extensions.