build: Use the new array type
meson 0.44 features a new option type called `array` that allows
more than one string to be passed.

This feature fits perfectly the `enable_extensions` option needs,
so it has been changed to be an `array` type. the option has not
been limited to a set of choices to avoid duplication.