auto-move-windows: Track apps' windows instead of window creation
We currently track window creation to decide whether the new window
should be moved or not. In order for this to work, the window must
already have been matched to the correct application, which is only
the case when the properties used for app matching were supplied
during window creation.

This is usually the case on X11, but never on wayland. To avoid this
issue altogether, stop listening for raw window creations, and instead
track when a window is added to an application we are interested in.

https://gitlab.gnome.org/GNOME/gnome-shell-extensions/issues/33