window-list: Temporarily revert back to Lang.Class
PanelMenu.Button was made a GObject subclass after the 3.30.1 release,
which means subclassing using ES6 classes works differently before and
after the release. But as extensions are expected to work for any point
release of a stable cycle, we cannot do that.

So temporarily revert back to Lang.Class which works in both cases.

https://gitlab.gnome.org/GNOME/gnome-shell-extensions/issues/113